Initial `install` command static worker support

+++ pkg/scripts/kubeadm.go
@@ -24,6 +24,13 @@ sudo kubeadm join {{ .VERBOSE }} \
 	--config=./{{ .WORK_DIR }}/cfg/master_{{ .NODE_ID }}.yaml
 `
 
+	kubeadmWorkerJoinScriptTemplate = `
+if [[ -f /etc/kubernetes/kubelet.conf ]]; then exit 0; fi
+
+sudo kubeadm join {{ .VERBOSE }} \
+--config=./{{ .WORK_DIR }}/cfg/worker_{{ .NODE_ID }}.yaml
+`
+
 	kubeadmCertScriptTemplate = `
 if [[ -d ./{{ .WORK_DIR }}/pki ]]; then
    sudo rsync -av ./{{ .WORK_DIR }}/pki/ /etc/kubernetes/pki/
@@ -52,7 +59,14 @@ rm -rf "{{ .WORK_DIR }}"
 sudo {{ .KUBEADM_UPGRADE }} --config=./{{ .WORK_DIR }}/cfg/master_0.yaml`
 )
 
-func KubeadmJoin(workdir string, nodeID int, verboseFlag string) (string, error) {
+func KubeadmJoin(workdir string, nodeID int, verboseFlag string, isWorker bool) (string, error) {
+	if isWorker {
+		return Render(kubeadmWorkerJoinScriptTemplate, Data{
+			"WORK_DIR": workdir,
+			"NODE_ID":  nodeID,
+			"VERBOSE":  verboseFlag,
+		})
+	}
 	return Render(kubeadmJoinScriptTemplate, Data{
 		"WORK_DIR": workdir,
 		"NODE_ID":  nodeID,
